PythonService/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
transformers==4.52.1
torch==2.9.0
soundfile==0.12.1
language-tool-python==2.8
Pillow==12.0.0
numpy==1.26.4
requests==2.32.3
openai==1.97.1
fastapi==0.115.2
uvicorn==0.32.0
pydantic~=2.11.7
protobuf~=4.25.8
librosa~=0.11.0
sentence-transformers~=5.0.0
PyJWT~=2.10.1
cryptography~=45.0.6
python-dotenv~=1.1.1
google-generativeai
asyncio
black
flake8
sqlalchemy
asyncpg
aioredis